refactor: split discord bot into discordbot.js
